Fixes security settings in scanned networks under the CC32xx. The API has changed from simplelink v1, resulting all found networks reported as "open".

+++ src/freertos_drivers/net_cc32xx/CC32xxWiFi.hxx
@@ -337,7 +337,7 @@ private:
      * @param sec_type simplelink security type
      * @return security type
      */
-    SecurityType security_type_from_simplelink(uint8_t sec_type);
+    SecurityType security_type_from_simplelink(unsigned sec_type);
 
     /** Set the CC32xx to its default state, including station mode.
      */
